Refactor propagator reconciler setup to one file
This commit only moves some code between files, it does change any
content.
