--- controllers/propagator/placementDecisionMapper.go
@@ -1,54 +1,0 @@
-// Copyright (c) 2020 Red Hat, Inc.
-// Copyright Contributors to the Open Cluster Management project
-
-package propagator
-
-import (
-	"context"
-
-	clusterv1beta1 "open-cluster-management.io/api/cluster/v1beta1"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/handler"
-	"sigs.k8s.io/controller-runtime/pkg/reconcile"
-
-	policiesv1 "open-cluster-management.io/governance-policy-propagator/api/v1"
-	"open-cluster-management.io/governance-policy-propagator/controllers/common"
-)
-
-func placementDecisionMapper(c client.Client) handler.MapFunc {
-	return func(ctx context.Context, object client.Object) []reconcile.Request {
-		log := log.WithValues("placementDecisionName", object.GetName(), "namespace", object.GetNamespace())
-
-		log.V(2).Info("Reconcile request for a placement decision")
-
-		// get the placement name from the placementdecision
-		placementName := object.GetLabels()["cluster.open-cluster-management.io/placement"]
-		if placementName == "" {
-			return nil
-		}
-
-		pbList := &policiesv1.PlacementBindingList{}
-		// find pb in the same namespace of placementrule
-		lopts := &client.ListOptions{Namespace: object.GetNamespace()}
-		opts := client.MatchingFields{"placementRef.name": placementName}
-		opts.ApplyToList(lopts)
-
-		err := c.List(ctx, pbList, lopts)
-		if err != nil {
-			return nil
-		}
-
-		var result []reconcile.Request
-		// loop through pbs and collect policies from each matching one.
-		for _, pb := range pbList.Items {
-			if pb.PlacementRef.APIGroup != clusterv1beta1.SchemeGroupVersion.Group ||
-				pb.PlacementRef.Kind != "Placement" || pb.PlacementRef.Name != placementName {
-				continue
-			}
-
-			result = append(result, common.GetPoliciesInPlacementBinding(ctx, c, &pb)...)
-		}
-
-		return result
-	}
-}
